# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EleevateOverseas Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: applicationId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload document
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document status
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Assessment ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        applicationId:
          type: string
        type:
          type: string
          enum: [sop, lor, transcript, passport, financial_proof, resume, english_test, other]
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, under_review]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, name, url, status, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        applicationId:
          type: string
        type:
          type: string
          enum: [sop, lor, transcript, passport, financial_proof, resume, english_test, other]
        name:
          type: string
        url:
          type: string
        notes:
          type: string
      required: [type, name, url]
...
    UpdateDocumentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, under_review]
        notes:
          type: string
...